A bit more general cleaning up
